Added additions, deletions, modifications, and checks

+++ model/alarmNum.go
@@ -6,5 +6,5 @@ type AlarmNum struct {
 	gorm.Model
 	IsAlarm   bool
 	Num       int64
-	InspectId int64
+	InspectId uint
 }
